+++ upcloud/provider.go
@@ -33,9 +33,6 @@ func Provider() *schema.Provider {
 			"upcloud_server":        resourceUpCloudServer(),
 			"upcloud_storage":       resourceUpCloudStorage(),
 			"upcloud_firewall_rule": resourceUpCloudFirewallRule(),
-			"upcloud_plan":          resourceUpCloudPlan(),
-			"upcloud_price":         resourceUpCloudPrice(),
-			"upcloud_price_zone":    resourceUpCloudPriceZone(),
 			"upcloud_tag":           resourceUpCloudTag(),
 		},