+++ CHANGELOG.md
@@ -17,4 +17,5 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
  
 ### Removed
  - Removed storage HCL blocks that failed due to referencing older UpCloud template ID
+ - Removed the plan, price, price_zone and timezone UpCloud resources